"GERMANY" Embalming Hager Werken +27787930326
Hager & Werken embalming products available in Johannesburg south Africa Embalming is the art and science of preserving human remains by treating them (in its modern form with chemicals) to forestall decomposition. The intention is to keep them suitable for public display at a funeral, for religious reasons, or for medical and scientific purposes such as their use as anatomical specimens. 
[1] The three goals of embalming are sensitization, presentation and preservation (or restoration). Embalming has a very long and cross-cultural history, with many cultures giving the embalming processes a greater religious meaning. Specific uses are: sensitization, presentation and preservation (or restoration). The type of embalming powder depends on the purity and color. Purity: 98%, 100% Hot Compound Origin:Germany Brand: Hager Werken PURITY ON COLOR. Pink 98% white 100% Distributor Dentist / Dental Technician Hager & Werken products are sold through the dental trade worldwide. Contact me on (+27) 787930326